+++ apps/backoffice-v2/src/common/components/organisms/TextEditor/components/ToolbarButton.tsx
@@ -1,0 +1,45 @@
+import * as React from 'react';
+import type { TooltipContentProps } from '@radix-ui/react-tooltip';
+import { Tooltip } from '@/common/components/atoms/Tooltip/Tooltip';
+import { TooltipTrigger } from '@/common/components/atoms/Tooltip/Tooltip.Trigger';
+import { TooltipContent } from '@/common/components/atoms/Tooltip/Tooltip.Content';
+import { Toggle } from '@/common/components/atoms/Toggle/Toggle';
+import { ctw } from '@ballerine/ui';
+
+interface ToolbarButtonProps extends React.ComponentPropsWithoutRef<typeof Toggle> {
+  isActive?: boolean;
+  tooltip?: string;
+  tooltipOptions?: TooltipContentProps;
+}
+
+export const ToolbarButton = React.forwardRef<HTMLButtonElement, ToolbarButtonProps>(
+  ({ isActive, children, tooltip, className, tooltipOptions, ...props }, ref) => {
+    const toggleButton = (
+      <Toggle
+        size="sm"
+        ref={ref}
+        className={ctw('size-8 p-0', { 'bg-accent': isActive }, className)}
+        {...props}
+      >
+        {children}
+      </Toggle>
+    );
+
+    if (!tooltip) {
+      return toggleButton;
+    }
+
+    return (
+      <Tooltip>
+        <TooltipTrigger asChild>{toggleButton}</TooltipTrigger>
+        <TooltipContent {...tooltipOptions}>
+          <div className="flex flex-col items-center text-center">{tooltip}</div>
+        </TooltipContent>
+      </Tooltip>
+    );
+  },
+);
+
+ToolbarButton.displayName = 'ToolbarButton';
+
+export default ToolbarButton;