+++ apps/kyb-app/src/hooks/useAppExit/useAppExit.ts
@@ -3,18 +3,20 @@ import { useFlowTracking } from '@/hooks/useFlowTracking';
 import { useLanguage } from '@/hooks/useLanguage';
 import { useUISchemasQuery } from '@/hooks/useUISchemasQuery';
 import { useCallback } from 'react';
+import { CollectionFlowEvents } from '../useFlowTracking/enums';
 
 export const useAppExit = () => {
   const appLanguage = useLanguage();
   const { data: uiSchema } = useUISchemasQuery(appLanguage);
-  const { trackExit } = useFlowTracking();
   const { customer } = useCustomerQuery();
+  const { trackEvent } = useFlowTracking();
 
   const kybOnExitAction = uiSchema?.config?.kybOnExitAction || 'send-event';
 
   const exit = useCallback(() => {
     if (kybOnExitAction === 'send-event') {
-      trackExit();
+      trackEvent(CollectionFlowEvents.USER_EXITED);
+
       return;
     }
 
@@ -23,11 +25,10 @@ export const useAppExit = () => {
         location.href = customer?.websiteUrl;
       }
     }
-  }, [trackExit, customer]);
+  }, [trackEvent, customer]);
 
   return {
     exit,
-    isExitAvailable:
-      uiSchema?.config?.kybOnExitAction === 'send-event' ? true : !!customer?.websiteUrl,
+    isExitAvailable: kybOnExitAction === 'send-event' ? true : !!customer?.websiteUrl,
   };
 };