+++ services/workflows-service/src/workflow/workflow-runtime-data.repository.intg.test.ts
@@ -31,13 +31,13 @@ import { UiDefinitionService } from '@/ui-definition/ui-definition.service';
 import { UiDefinitionRepository } from '@/ui-definition/ui-definition.repository';
 import { faker } from '@faker-js/faker';
 import { BusinessService } from '@/business/business.service';
-import { BusinessReportRepository } from '@/business-report/business-report.repository';
 import { BusinessReportService } from '@/business-report/business-report.service';
 import { RiskRuleService } from '@/rule-engine/risk-rule.service';
 import { RuleEngineService } from '@/rule-engine/rule-engine.service';
 import { NotionService } from '@/notion/notion.service';
 import { SentryService } from '@/sentry/sentry.service';
 import { SecretsManagerFactory } from '@/secrets-manager/secrets-manager.factory';
+import { MerchantMonitoringClient } from '@/business-report/merchant-monitoring-client';
 
 describe('#Workflow Runtime Repository Integration Tests', () => {
   let workflowRuntimeRepository: WorkflowRuntimeDataRepository;
@@ -60,7 +60,6 @@ describe('#Workflow Runtime Repository Integration Tests', () => {
       WorkflowEventEmitterService,
       BusinessRepository,
       BusinessReportService,
-      BusinessReportRepository,
       BusinessService,
       WorkflowDefinitionRepository,
       WorkflowService,
@@ -82,6 +81,7 @@ describe('#Workflow Runtime Repository Integration Tests', () => {
       NotionService,
       SentryService,
       SecretsManagerFactory,
+      MerchantMonitoringClient,
     ];
 
     workflowRuntimeRepository = (await fetchServiceFromModule(