+++ packages/react-pdf-toolkit/CHANGELOG.md
@@ -1,10 +1,35 @@
 # @ballerine/react-pdf-toolkit
 
-## 1.2.40
+## 1.2.43
 
 ### Patch Changes
 
 - Updated dependencies
+  - @ballerine/ui@0.5.43
+
+## 1.2.42
+
+### Patch Changes
+
+- Change
+- Updated dependencies
+  - @ballerine/config@1.1.23
+  - @ballerine/ui@0.5.42
+
+## 1.2.41
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/ui@0.5.41
+
+## 1.2.40
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/config@1.1.22
   - @ballerine/ui@0.5.40
 
 ## 1.2.39